added jupyter notebook to show plots of relevant raw counts and coverages from HHS data
