+++ .gitignore
@@ -1,4 +1,3 @@
-*.csv
 *.xlsx
 *.html
